+++ src/bda/plone/shop/shipping.py
@@ -1,4 +1,4 @@
-from bda.plone.cart import get_object_by_uid
+from bda.plone.cart import cart_item_shippable
 from bda.plone.cart.browser import CURRENCY_LITERALS
 from bda.plone.shipping import Shipping
 from bda.plone.shipping.interfaces import IShippingSettings
@@ -149,16 +149,19 @@ class DefaultShipping(Shipping):
         calc = CartItemCalculator(self.context)
         shipping_limit_from_gross = settings.shipping_limit_from_gross
         free_shipping_limit = Decimal(str(settings.free_shipping_limit))
+        purchase_price = Decimal(0)
         # calculate shipping from gross
         if shipping_limit_from_gross:
-            # XXX: consider shippable flag for items and calculate
-            #      purchase price from shippable items only
-            purchase_price = calc.net(items) + calc.vat(items)
+            for item in items:
+                if not cart_item_shippable(self.context, item):
+                    continue
+                purchase_price += calc.item_net(item) + calc.item_vat(item)
         # calculate shipping from net
         else:
-            # XXX: consider shippable flag for items and calculate
-            #      purchase price from shippable items only
-            purchase_price = calc.net(items)
+            for item in items:
+                if not cart_item_shippable(self.context, item):
+                    continue
+                purchase_price += calc.item_net(item)
         # purchase price exceeds free shipping limit, no shipping costs
         if free_shipping_limit and purchase_price > free_shipping_limit:
             return Decimal(0)
@@ -168,10 +171,7 @@ class DefaultShipping(Shipping):
         # item shipping costs set, calculate for contained cart items
         if item_shipping_cost > Decimal(0):
             for item in items:
-                obj = get_object_by_uid(self.context, item[0])
-                if not obj:
-                    continue
-                if IShippingItem(obj).shippable:
+                if not cart_item_shippable(self.context, item):
                     continue
                 shipping_costs += item_shipping_cost * item[1]
         # consider flat shipping cost if set
@@ -187,8 +187,6 @@ class DefaultShipping(Shipping):
     def vat(self, items):
         settings = get_shop_shipping_settings()
         shipping_vat = Decimal(str(settings.shipping_vat))
-        # XXX: consider shippable flag for items and calculate
-        #      vat from shippable items only
         return self.net(items) / Decimal(100) * shipping_vat
 
 
@@ -206,7 +204,12 @@ class FlatRate(Shipping):
 
     def calculate(self, items):
         calc = CartItemCalculator(self.context)
-        if calc.net(items) + calc.vat(items) > Decimal(FREE_SHIPPING_LIMIT):
+        purchase_price = Decimal(0)
+        for item in items:
+            if not cart_item_shippable(self.context, item):
+                continue
+            purchase_price += calc.item_net(item) + calc.item_vat(item)
+        if purchase_price > Decimal(FREE_SHIPPING_LIMIT):
             return Decimal(0)
         return Decimal(FLAT_SHIPPING_COST)